WIP: Writing tests for HttpConnection
