+++ tests/common/CMakeLists.txt
@@ -1,8 +1,15 @@
 add_library(clio_testing_common)
 
 target_sources(
-  clio_testing_common PRIVATE util/StringUtils.cpp util/TestHttpServer.cpp util/TestWsServer.cpp util/TestObject.cpp
-                              util/AssignRandomPort.cpp util/CallWithTimeout.cpp
+  clio_testing_common
+  PRIVATE util/AssignRandomPort.cpp
+          util/CallWithTimeout.cpp
+          util/StringUtils.cpp
+          util/TestHttpClient.cpp
+          util/TestHttpServer.cpp
+          util/TestObject.cpp
+          util/TestWebSocketClient.cpp
+          util/TestWsServer.cpp
 )
 
 include(deps/gtest)
